test(tcp): move shutdown worker into a typed fixture
